Move width calculation code to ViewUtil
